# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "SmartScheduler API",
    "description": "SmartScheduler API for job scheduling and contractor management",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5004"
    }
  ],
  "paths": {
...
    "/api/admin/weights": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Update weights configuration",
        "description": "Create a new version of the weights configuration",
        "operationId": "UpdateWeightsConfig",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWeightsConfigRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightsConfigResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/api/admin/weights/rollback": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Rollback weights configuration",
        "description": "Rollback to a previous version of the weights configuration",
        "operationId": "RollbackWeightsConfig",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackWeightsConfigRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightsConfigResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/activity": {
      "get": {
        "tags": [
          "Activity"
        ],
        "summary": "Get recent activities",
        "description": "Retrieves recent system activities from the event log. Supports filtering by activity types and limiting results.",
        "operationId": "GetActivities",
        "parameters": [
          {
            "name": "types",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "WeightsConfigResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "weights": {
            "$ref": "#/components/schemas/WeightFactorsDto"
          },
          "tieBreakers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rotation": {
            "$ref": "#/components/schemas/RotationConfigDto"
          },
          "changeNotes": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WeightFactorsDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "availability": {
            "type": "number",
            "format": "double"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "RotationConfigDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "boost": {
            "type": "number",
            "format": "double"
          },
          "underUtilizationThreshold": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "UpdateWeightsConfigRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "weights": {
            "$ref": "#/components/schemas/WeightFactorsDto"
          },
          "tieBreakers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rotation": {
            "$ref": "#/components/schemas/RotationConfigDto"
          },
          "changeNotes": {
            "type": "string"
          }
        }
      },
      "RollbackWeightsConfigRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "changeNotes": {
            "type": "string"
          }
        }
      },
...
      "ActivityDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      }
    }
  }
}